requirements.txt
Flask==2.1.2
requests==2.28.1
PyYAML==6.0
gunicorn==20.1.0
redis==4.3.4
sqlalchemy==1.4.36
psycopg2-binary==2.9.3
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
openai==0.27.0